experiments/requirements.txt
email-validator==2.2.0
gradio==4.21.0
httptools==0.6.4
importlib-metadata==8.5.0
ipykernel==6.29.5
itsdangerous==2.2.0
jaraco.collections==5.1.0
loguru==0.7.0
pickleshare==0.7.5
pip-chill==1.0.3
pydantic-extra-types==2.9.0
pydantic-settings==2.6.0
pyfiglet==1.0.2
roboflow==1.1.47
toml==0.10.2
tomli==2.0.1
ujson==5.10.0
ultralytics==8.3.12
uvloop==0.21.0
watchfiles==0.24.0
